+++ frontend/src/pages/AnnouncementPage.js
@@ -6,275 +6,18 @@ import React, {
 	useState
 } from "react";
 import appwriteApi from "../api/appwriteApi";
-import useChangeRoute from "../hooks/useChangeRoute";
 import { Link as RouterLink, useLocation } from "react-router-dom";
-
-import Grid from "@mui/material/Grid";
-import { Button, TextField, Typography } from "@mui/material";
+import { Button } from "@mui/material";
 import Box from "@mui/material/Box";
-import Collapse from "@mui/material/Collapse";
-import Link from "@mui/material/Link";
 
 import ConditionalAlert from "../components/ConditionalAlert";
 import { useTeamMembership } from "../components/RestrictedArea";
 import RoundedEdgesButton from "../components/RoundedEdgesButton";
 import HeaderTypography from "../components/HeaderTypography";
-import ParagraphTypography from "../components/ParagraphTypography";
-
-import examplePhoto1 from "../assets/img/announcementPhotoExamples/anouncement_1.png";
-import examplePhoto2 from "../assets/img/announcementPhotoExamples/anouncement_2.png";
-import examplePhoto3 from "../assets/img/announcementPhotoExamples/anouncement_3.png";
-import examplePhoto4 from "../assets/img/announcementPhotoExamples/anouncement_4.png";
-import examplePhoto5 from "../assets/img/announcementPhotoExamples/anouncement_5.png";
-import {
-	semiTransparentTextColor,
-	textColor,
-	whiteBoxBorderColor,
-	whiteTransparentBorderColor
-} from "../assets/jss/colorPalette";
-import { headerFont, linkFont, textFont } from "../assets/jss/fontPalette";
-
-const photos = [examplePhoto1, examplePhoto2, examplePhoto3, examplePhoto4, examplePhoto5];
-
-function InputFields({ defaultTitle, defaultContent, titleComponenId, contentComponenId }) {
-	// TODO: formated text support
-	// ref: https://mui.com/components/text-fields/#integration-with-3rd-party-input-libraries
-	return <Grid sx={{ mt: 3, marginBottom: 3 }}>
-		<Grid item xs={12}>
-			<TextField
-				required
-				fullWidth
-				name="title"
-				label="Title"
-				id={titleComponenId}
-				defaultValue={defaultTitle}
-				multiline
-				minRows={1}
-				maxRows={10}
-				sx={{ mt: 1, marginBottom: 1 }}
-			/>
-		</Grid>
-		<Grid item xs={12}>
-			<TextField
-				required
-				fullWidth
-				name="content"
-				label="Content"
-				id={contentComponenId}
-				defaultValue={defaultContent}
-				multiline
-				minRows={1}
-				maxRows={10}
-				sx={{ mt: 1, marginBottom: 1 }}
-			/>
-		</Grid>
-	</Grid>;
-}
-
-function AnnouncementEntry({
-	announcement, editing, setEditing, userIsAdmin, setAnnouncementsAreUpToDate, isSidebar
-}) {
-
-	const handleEditButton = id => () => {
-		setEditing(id);
-	};
-
-	const changeRoute = useChangeRoute();
-
-	const handleDeleteButton = id => () => {
-		appwriteApi.deleteAnnouncement(id)
-			.then(() => {
-				setAnnouncementsAreUpToDate(false);
-			})
-			.catch((e) => {
-				console.log(e);
-			});
-	};
-
-	const handleCancelButton = () => {
-		setEditing("");
-		if (window.location.hash.split("#")[1]) {
-			changeRoute("/announcements");
-		}
-	};
-
-	const handleSubmitButton = (announcementId, titleComponenId, contentComponenId) => () => {
-		const title = document.getElementById(titleComponenId);
-		const content = document.getElementById(contentComponenId);
-		if (title.value.length == 0 || content.value.length == 0) {
-			console.log("missing input or content!");
-			return;
-		}
-		// Note: js get Unix time in Milisecond. Backend uses Python which utilizes *second, 
-		// so for consistency we store all time in seconds 
-		appwriteApi.updateAnnouncement({
-			"title": title.value,
-			"content": content.value,
-			"updated_at": new Date().valueOf() / 1000 | 0
-		}, announcementId)
-			.then(() => {
-				setAnnouncementsAreUpToDate(false);
-				if (window.location.hash.split("#")[1]) {
-					changeRoute("/announcements");
-				}
-			})
-			.catch((e) => {
-				console.log(e);
-			});
-	};
-
-	/* Prepare datetime */
-	const created_at = new Date(announcement.created_at * 1000);
-	const formated_created_at = created_at.toString().substring(4,16);
-
-	const limitLines = isSidebar ? {
-		display: "-webkit-box",
-		overflow: "hidden",
-		WebkitBoxOrient: "vertical",
-		WebkitLineClamp: 2,
-	} : {};
-
-	const boxPageStyle = { display: "flex", mt: 1, marginBottom: 3, p: 1, border: 1, borderColor: whiteTransparentBorderColor };
-	const titleStyle = { fontFamily: headerFont, fontSize: "14px", fontStyle: "normal", fontWeight: "bold" };
-	const dateStyle = {
-		fontFamily: textFont, fontSize: "11px", fontStyle: "normal",
-		fontWeight: "medium", opacity: 0.4
-	};
-	const contentStyle = {
-		fontFamily: textFont, fontSize: "12px",
-		fontStyle: "normal", fontWeight: "medium",
-		color: semiTransparentTextColor
-	};
-	const linkStyle = {
-		fontFamily: linkFont, fontSize: "12px",
-		fontStyle: "normal", fontWeight: "medium",
-		color: textColor, textDecoration: "underline",
-	};
-	const announcePhoto = photos[parseInt(announcement.$id, 16) % photos.length];
-
-	const adminControls = userIsAdmin && <>
-		|&nbsp;&nbsp;
-		<Typography onClick={handleDeleteButton(announcement.$id)} style={linkStyle}>
-			Delete &#x2715;
-		</Typography>
-		&nbsp; | &nbsp;
-		<RouterLink to={"/announcements#" + announcement.$id} style={linkStyle} >
-			Edit &#9881;
-		</RouterLink>
-	</>;
-
-	const sidebarComponents =
-		<div style={{ marginBottom: "2px" }}>
-			<Box>
-				<Grid container spacing={0} sx={{ maxWidth: "sm" }}>
-					<Grid item xs={9}>
-						<Link href="/announcements" style={{}}>
-							<div style={{ backgroundImage: `url(${announcePhoto})`, backgroundSize: "cover", backgroundRepeat: "no-repeat", backgroundPosition: "center", width: "126px", height: "100px", borderRadius: "10px", }} />
-						</Link>
-					</Grid>
-					<Grid
-						item xs
-						sx={{
-							minHeight: "100px", marginLeft: "8px", paddingLeft: 0, paddingBottom: "10px", borderBottom: 1,
-							borderColor: whiteTransparentBorderColor, marginBottom: 2,
-						}}
-					>
-						<HeaderTypography style={titleStyle} sx={limitLines} variant="h5">
-							{announcement.title}
-						</HeaderTypography>
-						<Box style={{ display: "flex", flexDirection: "row", marginBottom: 1, marginTop: 5 }}>
-							<Typography style={dateStyle} sx={{ marginBottom: 1 }}>
-								{formated_created_at}&nbsp;&nbsp;&nbsp;&nbsp;
-							</Typography>
-							{adminControls}
-						</Box>
-						<ParagraphTypography style={contentStyle} sx={limitLines}>
-							{announcement.content}
-						</ParagraphTypography>
-						<Typography component={RouterLink} to={`/announcements#${announcement.$id}`} style={linkStyle}>Read more</Typography>
-					</Grid>
-				</Grid>
-			</Box>
-		</div>;
-	
-	if (isSidebar)
-		return sidebarComponents;
-
-	const pageComponents =
-		<div style={{ width: "100%" }}>
-			<Box xs={12} sx={boxPageStyle}>
-				<div style={{ width: "100%", padding: 5 }}>
-					<Typography style={titleStyle} sx={limitLines} variant="h5">
-						{announcement.title}
-					</Typography>
-					<Typography style={dateStyle} sx={{ marginBottom: 1 }}>{formated_created_at}</Typography>
-					<Typography style={contentStyle} sx={{ marginBottom: 1 , ...limitLines }}>
-						{announcement.content}
-					</Typography>
-					{userIsAdmin &&
-						<div style={{ textAlign: "center", marginTop: 3 }}>
-							<Button onClick={handleDeleteButton(announcement.$id)} variant="outlined" sx={{ m: 1 }}>
-								Delete
-							</Button>
-							<Button onClick={handleEditButton(announcement.$id)} variant="outlined" sx={{ m: 1 }}>
-								Edit
-							</Button>
-						</div>
-					}
-				</div>
-			</Box>
-			{/* Add edit Collapse component for each Announcement entry if user is admin */}
-			{userIsAdmin &&
-				<Collapse in={editing == announcement.$id ? true : false}>
-					<Box sx={{ m: 2, p: 2, backgroundColor: whiteBoxBorderColor, borderRadius: "15px" }}>
-						<InputFields
-							defaultTitle={announcement.title}
-							defaultContent={announcement.content}
-							titleComponenId={"edit_title_" + announcement.$id}
-							contentComponenId={"edit_content_" + announcement.$id}
-						/>
-						<div style={{ textAlign: "center" }}>
-							<Button
-								onClick={handleSubmitButton(
-									announcement.$id,
-									"edit_title_" + announcement.$id,
-									"edit_content_" + announcement.$id
-								)}
-								variant="contained" sx={{ ma: 2 }}
-							>
-								Submit
-							</Button>
-							<Button onClick={handleCancelButton} variant="contained" sx={{ m: 2 }}>
-								Cancel
-							</Button>
-						</div>
-					</Box>
-				</Collapse>
-			}
-		</div>;
-
-	return pageComponents;
-}
-
-function AnnouncementContainer({
-	announcements, editing, setEditing, userIsAdmin, setAnnouncementsAreUpToDate, isSidebar
-}) {
-	return <div>
-		{announcements.map((announcement, index) => {
-			announcement["index"] = index;
-			return <div id={"c" + announcement.$id} key={announcement.$id}>
-				<AnnouncementEntry
-					announcement={announcement}
-					editing={editing}
-					setEditing={setEditing}
-					userIsAdmin={userIsAdmin}
-					setAnnouncementsAreUpToDate={setAnnouncementsAreUpToDate}
-					isSidebar={isSidebar}
-				/>
-			</div>;
-		})}
-	</div>;
-}
+import { whiteBoxBorderColor } from "../assets/jss/colorPalette";
+import { headerFont } from "../assets/jss/fontPalette";
+import InputFields from "../components/InputFields";
+import AnnouncementContainer from "../components/AnnouncementContainer";
 
 /**
  * Component to view announcements.
@@ -389,8 +132,8 @@ export default function AnnouncementPage(user, isSidebar) {
 	return <div component="main" style={fullWidth}>
 		{!isSidebar && userIsAdmin && <>
 			<AddAnnouncement />
-			<ConditionalAlert severity="error" text={errorMessageAddAnnouncement}/>
-			<ConditionalAlert severity="error" text={errorMessageGetAnnouncement}/>
+			<ConditionalAlert severity="error" text={errorMessageAddAnnouncement} />
+			<ConditionalAlert severity="error" text={errorMessageGetAnnouncement} />
 		</>}
 		<Box sx={{ ml: 4, p: 0 }}>
 			{isSidebar
